Move SetupJdkTask to task util package.
