+++ src/main/java/net/javasauce/ss/tasks/util/package-info.java
@@ -1,0 +1,4 @@
+@NonNullApi
+package net.javasauce.ss.tasks.util;
+
+import net.covers1624.quack.annotation.NonNullApi;
